Masterboa

+++ src/Shared/Kernel/Responses/SpResultBase.cs
@@ -4,6 +4,6 @@
     {
         public bool Success { get; set; }
         public string Message { get; set; } = string.Empty;
-        public long Data { get; set; }
+        public int Data { get; set; }
     }
 }
